requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
gevent==23.9.1
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
Werkzeug==3.0.0
zope.event==5.0
zope.interface==6.1
fdb
jsonify
mgrs
flask_restful
flask_cors
paho-mqtt==2.1.0
owlready2
wtforms
requests
flask_wtf

